Work around ActiveResource caching issue
When ActiveResource loads the asset JSON, it creates subclasses such
as Asset::Sizes on-the-fly. When we load an asset from cache, that
doesn't exist and the load fails.  This workaround uses a class
variable to say that the first asset an instance loads needs to
come from the server and not from cache.