Exclude calling content item from category carousel via sphinx
NOTE: obj_key is stored in Sphinx via a CRC32 checksum. There is a non-zero
chance of collision in checksums, meaning that we could accidentally exclude
something we don't mean to exclude.  Given the low probability and the low
cost if it does happen (something else gets displayed instead), this is ok
for our purposes.